refactor(hooks): split useConvocations into focused domain hooks
Extract 839-line monolithic hook into focused modules for improved
maintainability and easier testing:

- usePaginatedQuery.ts: Reusable pagination utilities and constants
- useAssignments.ts: Assignment queries (upcoming, past, validation-closed)
- useCompensations.ts: Compensation queries and mutations
- useExchanges.ts: Exchange queries and mutations
- useSettings.ts: Association settings and season hooks

The original useConvocations.ts is now a barrel file that re-exports
all hooks for backwards compatibility.
