+++ web-app/src/hooks/useSettings.ts
@@ -1,0 +1,41 @@
+import { useQuery, type UseQueryResult } from "@tanstack/react-query";
+import { api, type AssociationSettings, type Season } from "@/api/client";
+import { useAuthStore } from "@/stores/auth";
+import { queryKeys } from "@/api/queryKeys";
+
+/**
+ * Hook to fetch association settings.
+ * Settings include deadline hours for validation and other configuration.
+ *
+ * Note: Disabled in demo mode as demo data doesn't need these settings.
+ */
+export function useAssociationSettings(): UseQueryResult<
+  AssociationSettings,
+  Error
+> {
+  const isDemoMode = useAuthStore((state) => state.isDemoMode);
+
+  return useQuery({
+    queryKey: queryKeys.settings.association(),
+    queryFn: () => api.getAssociationSettings(),
+    staleTime: 30 * 60 * 1000, // 30 minutes - settings rarely change
+    enabled: !isDemoMode,
+  });
+}
+
+/**
+ * Hook to fetch the currently active season.
+ * Used to determine date ranges for assignment queries.
+ *
+ * Note: Disabled in demo mode as demo data uses fixed date ranges.
+ */
+export function useActiveSeason(): UseQueryResult<Season, Error> {
+  const isDemoMode = useAuthStore((state) => state.isDemoMode);
+
+  return useQuery({
+    queryKey: queryKeys.seasons.active(),
+    queryFn: () => api.getActiveSeason(),
+    staleTime: 60 * 60 * 1000, // 1 hour - season rarely changes
+    enabled: !isDemoMode,
+  });
+}
